refactor(approval-audit): Split routes out of controllers
approval-service's approvalRequest.controller.ts and
quick-response.controller.ts, and audit-service's auditLog.controller.ts
move route/middleware declarations and response Zod schemas into their
own *.routes.ts files; controllers keep only request-handling logic via
createXController(service) factories. module.ts files now create the
shared DocumentedRouter and register routes into it.

No behavior change — verified route-for-route identical OpenAPI doc
before/after for both services, lint clean, all 52 (approval) and 6
(audit) existing tests pass, and live smoke checks against the running
services.
